+++ drivers/infiniband/hw/ocrdma/ocrdma_ah.c
@@ -250,7 +250,7 @@ int ocrdma_query_ah(struct ib_ah *ibah, struct rdma_ah_attr *attr)
 
 int ocrdma_process_mad(struct ib_device *ibdev,
 		       int process_mad_flags,
-		       u8 port_num,
+		       u32 port_num,
 		       const struct ib_wc *in_wc,
 		       const struct ib_grh *in_grh,
 		       const struct ib_mad_hdr *in, size_t in_mad_size,
